src/models-api/requirements.txt
Flask==2.2.2
flask_restful==0.3.9
python-decouple==3.7
more-itertools
ffmpeg-python==0.2.0
transformers>=4.19.0
tqdm
setuptools-rust
numpy
torch
gdown
spacy
pyspellchecker
deep_translator
nltk
scikit-learn 
scipy 
matplotlib